% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dists}
\alias{dists}
\alias{is.dists}
\title{Class dists: Collection of Distances and Ratios}
\usage{
dists(df, area, n_obj, max_dist)

is.dists(obj)
}
\arguments{
\item{df}{A data frame with the columns \code{sim}, \code{dist}, and \code{ratio}
containing an indicator of the model run (0:n_sim), distances between
the objects of the patterns, and the ratios of a buffer with distance
\code{dist} inside the study area (needed for Ripley's edge correction).}

\item{area}{size of the study area in square units}

\item{n_obj}{number of objects in the pattern}

\item{max_dist}{maximum distance to be measured in pattern}

\item{obj}{an R object, preferably of class \code{dists}}
}
\value{
An object of class \code{dists}.
}
\description{
Advanced Use Only. This low-level function creates an object of class
"dists" from raw numerical data.
}
