% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_read_wkb.R
\name{test_reading_polys}
\alias{test_reading_polys}
\title{Test old Reading}
\usage{
test_reading_polys(dsn)
}
\arguments{
\item{dsn}{name of a data source}
}
\value{
nothing but the C++ function prints the WKT of the geometry to stdout
}
\description{
Test old Reading
}
\examples{
test_reading_polys(system.file("shapes/sim_area.shp", package="apcf"))
test_reading_polys(system.file("shapes/sim_pat_reg.shp", package="apcf"))

}
