% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_read_wkb.R
\name{test_reading_wkb}
\alias{test_reading_wkb}
\title{Test WKB Reading}
\usage{
test_reading_wkb(wkb)
}
\arguments{
\item{wkb}{a list of wkb}
}
\value{
nothing but the C++ function prints the WKT of the geometry to stdout
}
\description{
Test WKB Reading
}
\examples{
test_reading_wkb(sim_area_wkb)
test_reading_wkb(sim_pat_reg_wkb)

}
