% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_read_wkb.R
\name{test_rw_wkb}
\alias{test_rw_wkb}
\title{Test read and write WKB}
\usage{
test_rw_wkb(wkb)
}
\arguments{
\item{wkb}{an R Object}
}
\value{
list of raw (wkb)
}
\description{
shall read wkb from R Object and write WKB To file
}
\examples{
wkb_a <- test_rw_wkb(sim_area_wkb)
wkb_p <- test_rw_wkb(sim_pat_reg_wkb)
}
